fix codegen commands
